pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=62.3"]


[project]
authors = [
  {name = "Jordy", email = "jordyfaria0@gmail.com"},
]
classifiers = []
dependencies = [
	"python-dotenv==1.0.1"
]
# description = "Python Matter WebSocket Server"
# license = {text = "Apache-2.0"}
name = "teste"
readme = "README.md"
requires-python = ">=3.11"
# The version is set by GH action on release!
version = "0.0.0"


[project.optional-dependencies]
dev = [
	"colorama==0.4.6"
]